Add spend accounting + MCP agent surface
- pricing/: parse usage (OpenAI + Anthropic, streaming + non-streaming),
  price per model, fall back to a non-zero charge when usage is missing so
  usage-less calls can't dodge the cap (eng-review tripwire closed)
- proxy: tee LLM responses — stream one copy to the grantee, drain the other
  to compute cost, charge the grant + record per-call cost on the audit row
- store: spent_cents/est_cost_cents -> REAL (fractional cents); writeAudit
  returns id; add updateAuditCost
- mcp/: real stdio MCP server, a thin client of the proxy exposing a `request`
  tool; agent holds only a revocable grant token, never the key

54 tests passing (+ pricing, meter, forward); typecheck clean.
MCP initialize handshake verified.

+++ src/store/db.ts
@@ -60,7 +60,7 @@ export function initSchema(raw: Database.Database): void {
       token_hash      TEXT NOT NULL UNIQUE,
       scope_json      TEXT NOT NULL,
       spend_cap_cents INTEGER,
-      spent_cents     INTEGER NOT NULL DEFAULT 0,
+      spent_cents     REAL NOT NULL DEFAULT 0,
       expires_at      INTEGER,
       revoked_at      INTEGER,
       created_at      INTEGER NOT NULL
@@ -77,7 +77,7 @@ export function initSchema(raw: Database.Database): void {
       upstream_status INTEGER,
       bytes_in        INTEGER,
       bytes_out       INTEGER,
-      est_cost_cents  INTEGER
+      est_cost_cents  REAL
     );
     CREATE INDEX IF NOT EXISTS idx_audit_grant ON audit(grant_id);
   `);
